Check for -1 hashes after conversion via c_long
This ensures the hash value has its final sign so that the comparison
works as expected.

For example, in shash(), consider that masking -1, (-1 &
0xffffffffffffffff), gives the positive value 0xffffffffffffffff; but
ctypes.c_long(0xffffffffffffffff).value interprets it as a 64-bit
two’s-complement bit pattern, giving -1 again.

Unfortunately, adding a test for this is not easy, since the space of
hashes is too large for brute-force search to be practical. Finding a
test input would require using mathematical techniques to shrink the
search space by working backwards from the hash value -1, which is
probably possible for this hash algorithm but is not quite trivial.